Add config parameter for step size, Use integers for faster encoders
A new parameter 's' can be set for the size of each step from the step/direction input. e.g. 10s changes the setpoint by 10 positions each step. 1s is a one to one step to position setting. Parameters will need to be reset after this update. e.g. you will loose you p.i.d. settings. 

The initial versions just used doubles for everything and that was really slow in the interrupt routine for the encoder and step/direction inputs. This code moves to integers (longs... 32 bits -2,147,483,648 to +2,147,483,647) for the position and setpoint which should support much faster rotation and higher resolution encoders.